Move concurrent instrumentation utils out from javaagent-instrumentation-api
