+++ instrumentation/executors/bootstrap/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+  id("otel.javaagent-bootstrap")
+}
