+++ instrumentation/executors/bootstrap/src/main/java/io/opentelemetry/javaagent/bootstrap/concurrent/TaskAdviceHelper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.api.concurrent;
+package io.opentelemetry.javaagent.bootstrap.concurrent;
 
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
