+++ instrumentation/okhttp/okhttp-3.0/javaagent/build.gradle.kts
@@ -11,16 +11,9 @@ muzzle {
   }
 }
 
-/*
-Note: there is a bit of dependency exclusion magic going on.
-We have to exclude all transitive dependencies on 'okhttp' because we would like to force
-specific version. We cannot use . Unfortunately we cannot just force version on
-a dependency because this doesn't work well with version ranges - it doesn't select latest.
-And we cannot use configurations to exclude this dependency from everywhere in one go
-because it looks like exclusions using configurations excludes dependency even if it explicit
-not transitive.
- */
 dependencies {
+  bootstrap(project(":instrumentation:executors:bootstrap"))
+
   implementation(project(":instrumentation:okhttp:okhttp-3.0:library"))
 
   library("com.squareup.okhttp3:okhttp:3.0.0")
