+++ instrumentation/undertow-1.4/javaagent/build.gradle.kts
@@ -14,6 +14,7 @@ muzzle {
 dependencies {
   library("io.undertow:undertow-core:2.0.0.Final")
 
+  bootstrap(project(":instrumentation:executors:bootstrap"))
   bootstrap(project(":instrumentation:servlet:servlet-common:bootstrap"))
   bootstrap(project(":instrumentation:undertow-1.4:bootstrap"))
 }
